+++ graph_pkg/edit_cost/edit_cost_letter.pxd
@@ -8,6 +8,6 @@ cdef class EditCostLetter(EditCost):
         double x1, y1, x2, y2, result
         int valence
 
-    cdef float _compute_cost_insert_node(self, float x1, float y1) except? -1
+    cdef double _compute_cost_insert_node(self, double x1, double y1) except? -1
 
     cdef double _compute_cost_substitute_node(self, double x1, double y1, double x2, double y2) except? -1
